ref: move settings creation to sentry init command
