+++ src/sentry/runner/commands/init.py
@@ -1,6 +1,36 @@
+import importlib.resources
 import os
 
 import click
+from django.utils.crypto import get_random_string
+
+from sentry.runner.settings import DEFAULT_SETTINGS_CONF, DEFAULT_SETTINGS_OVERRIDE
+
+
+def generate_secret_key() -> str:
+    chars = "abcdefghijklmnopqrstuvwxyz0123456789!@#%^&*(-_=+)"
+    return get_random_string(50, chars)
+
+
+def _load_config_template(path: str, version: str = "default") -> str:
+    return importlib.resources.files("sentry").joinpath(f"data/config/{path}.{version}").read_text()
+
+
+def _generate_settings(dev: bool = False) -> tuple[str, str]:
+    """
+    This command is run when ``default_path`` doesn't exist, or ``init`` is
+    run and returns a string representing the default data to put into their
+    settings file.
+    """
+    context = {
+        "secret_key": generate_secret_key(),
+        "debug_flag": dev,
+        "mail.backend": "console" if dev else "smtp",
+    }
+
+    py = _load_config_template(DEFAULT_SETTINGS_OVERRIDE, "default") % context
+    yaml = _load_config_template(DEFAULT_SETTINGS_CONF, "default") % context
+    return py, yaml
 
 
 @click.command()
@@ -13,7 +43,7 @@ import click
 @click.argument("directory", required=False)
 def init(dev: bool, no_clobber: bool, directory: str) -> None:
     "Initialize new configuration directory."
-    from sentry.runner.settings import discover_configs, generate_settings
+    from sentry.runner.settings import discover_configs
 
     if directory is not None:
         os.environ["SENTRY_CONF"] = directory
@@ -33,7 +63,7 @@ def init(dev: bool, no_clobber: bool, directory: str) -> None:
 
     os.makedirs(directory, exist_ok=True)
 
-    py_contents, yaml_contents = generate_settings(dev)
+    py_contents, yaml_contents = _generate_settings(dev)
 
     if not os.path.isfile(yaml):
         write_yaml = True
